feat: add in-header nav links for md+
