+++ components/Header/Logo.tsx
@@ -3,7 +3,7 @@ import Link from "next/link"
 
 export const Logo = () => {
   return (
-    <Typography variant="h6" sx={{ flexGrow: 1 }}>
+    <Typography variant="h6">
       <Link href="/">
         <a style={{ textDecoration: "none", color: "inherit" }}>Guilds</a>
       </Link>
